requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
drf-yasg==1.21.7
